+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added command `list-owners` that lists all owners specified in the CODEOWNERS file
+
 ### Changed
 - Renamed `\CodeOwners\Cli\Tests\Command\ListCommandTest` to `\CodeOwners\Cli\Tests\Command\ListFilesCommandTest`
+- Updated required version of `timoschinkel/codeowners` to `^1.1.0`
 
 ## [1.0.0] - 2020-01-08
 ### Added